requirements.txt
amqp==1.4.9
anyjson==0.3.3
asgi-redis==1.4.3
asgiref==1.1.2
asn1crypto==0.24.0
attrs==17.4.0
autobahn==17.10.1
Automat==0.6.0
Babel==2.5.1
bcrypt==3.1.4
billiard==3.3.0.23
cachetools==2.0.1
celery==3.1.25
certifi==2017.11.5
cffi==1.11.2
channels==1.1.8
chardet==3.0.4
constantly==15.1.0
cryptocompare==0.3
cryptography==2.1.4
daphne==1.4.2
Django==1.11.9
-e git+https://github.com/limpbrains/django-cc.git@80cd0027790bc7e3319736bc57c6c54849e16419#egg=django_cc
django-extra-views==0.6.4
django-haystack==2.6.1
django-oscar==1.5.1
django-phonenumber-field==1.3.0
django-qrcode==0.3
django-redis==4.8.0
django-tables2==1.17.1
django-treebeard==4.2.0
django-widget-tweaks==1.4.1
dj-database-url==0.4.2
Fabric3==1.13.1.post1
factory-boy==2.9.2
Faker==0.8.8
flake8==3.5.0
google-api-core==0.1.4
google-auth==1.3.0
googleapis-common-protos==1.5.3
grpcio==1.8.4
grpcio-tools==1.8.4
honcho==1.0.1
hyperlink==17.3.1
idna==2.6
incremental==17.5.0
kombu==3.0.37
mccabe==0.6.1
mock==2.0.0
msgpack-python==0.4.8
paramiko==2.4.0
pbr==3.1.1
phonenumberslite==8.8.8
Pillow==5.0.0
protobuf==3.5.1
purl==1.3.1
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycodestyle==2.3.1
pycoin==0.80
pycountry==17.9.23
pycparser==2.18
pyflakes==1.6.0
PyNaCl==1.2.1
python-bitcoinrpc==1.0
python-dateutil==2.6.1
python-decouple==3.1
pytz==2017.3
psycopg2==2.7.3.2
redis==2.10.6
requests==2.18.4
rsa==3.4.2
six==1.11.0
sorl-thumbnail==12.4.1
text-unidecode==1.1
Twisted==17.9.0
txaio==2.8.2
Unidecode==0.4.21
urllib3==1.22
whitenoise==3.3.1
zope.interface==4.4.3
